Add files via upload
Part 1.

Regarding dataset investigation (gender, age, job respective assessment of personal mood & behavior, job outlook assessment of personal attitude to life), correlation analysis (PCA), Hierarchical clustering, 
An overall conclusion on variables affecting job rate satisfaction (Q90f):
•	Respondents are confident in their respective roles (Q90f) not standing with the varying responses in emotional well-being questions. 
•	Positive and neutral-negative responses translate to confidence in work performance. 
•	Confidence in work abilities (Q90f) is not exclusively due to positive emotional well-being responses.

Part 2.

Worked with Portuguese dataset.
- translated categorical variables into binary via one-hot encoding & label-encoding algorithms
- running linear regressions for both datasets has shown that many variables do not need to be interpreted as there are too many of them
- plotted graphs of residuals against fitted to demonstrate that it is tough to make a prediction for both of the built regressions
- conducted stepwise (backward) regression as it will remove variables at each step from the regression model to figure out a reduced model that explains at its best the data provided
Conclusion:
The problem of multicollinearity in this task is of great importance especially for the model result and should be considered by all means as we were provided with such a great number of parameters. In this task, I tried to predict the final grade not taking into account G1 and G2 grades by using the multiple linear regression.

Part 3.

A problem of customers subscribing to a bank term deposit.
- Decision tree classifier
- Logistic regression (best at classifying samples) 
- Gradient boosting classifier
Conclusion:
The Logistic regression turned out to be one of the best performing models with 79.2 % accuracy, i.e. the Logistic regression model is able to predict with 79.2% accuracy that customers will subscribe to a bank term deposit.